+++ pylette/src/extractors/oklab.py
@@ -27,61 +27,10 @@ from numpy.typing import NDArray
 from typing_extensions import override
 
 from pylette.src.color import Color
+from pylette.src.colorspaces import linear_srgb_to_oklab, linear_to_srgb, oklab_to_linear_srgb, srgb_to_linear
 from pylette.src.extractors.protocol import NP_T, ColorExtractorBase
 from pylette.src.extractors.registry import register
-from pylette.src.types import ExtractionMethod, FloatArray
-
-
-# sRGB <-> linear sRGB (IEC 61966-2-1 transfer function)
-def srgb_to_linear(srgb: FloatArray) -> FloatArray:
-    """Map gamma-encoded sRGB in ``[0, 1]`` to linear-light sRGB in ``[0, 1]``."""
-    return np.where(srgb <= 0.04045, srgb / 12.92, ((srgb + 0.055) / 1.055) ** 2.4)
-
-
-def linear_to_srgb(linear: FloatArray) -> FloatArray:
-    """Map linear-light sRGB in ``[0, 1]`` back to gamma-encoded sRGB in ``[0, 1]``."""
-    linear = np.clip(linear, 0.0, 1.0)
-    return np.where(linear <= 0.0031308, 12.92 * linear, 1.055 * np.power(linear, 1.0 / 2.4) - 0.055)
-
-
-# linear sRGB <-> OKLab (Ottosson 2020)
-#
-# Forward matrices are the ones from the webpage. the inverse matrices are derived
-# from them so the round-trip is numerically self-consistent (single source of
-# truth) rather than two independently transcribed constant sets.
-
-# linear sRGB -> LMS
-_LRGB_TO_LMS = np.array(
-    [
-        [0.4122214708, 0.5363325363, 0.0514459929],
-        [0.2119034982, 0.6806995451, 0.1073969566],
-        [0.0883024619, 0.2817188376, 0.6299787005],
-    ]
-)
-# nonlinear l'm's' -> OKLab
-_LMS_TO_OKLAB = np.array(
-    [
-        [0.2104542553, 0.7936177850, -0.0040720468],
-        [1.9779984951, -2.4285922050, 0.4505937099],
-        [0.0259040371, 0.7827717662, -0.8086757660],
-    ]
-)
-_OKLAB_TO_LMS = np.linalg.inv(_LMS_TO_OKLAB)
-_LMS_TO_LRGB = np.linalg.inv(_LRGB_TO_LMS)
-
-
-def linear_srgb_to_oklab(rgb: FloatArray) -> FloatArray:
-    """Convert an ``(N, 3)`` array of linear sRGB to OKLab."""
-    lms = rgb @ _LRGB_TO_LMS.T
-    lms_nonlinear = np.cbrt(lms)
-    return lms_nonlinear @ _LMS_TO_OKLAB.T
-
-
-def oklab_to_linear_srgb(lab: FloatArray) -> FloatArray:
-    """Convert an ``(N, 3)`` array of OKLab back to linear sRGB."""
-    lms_nonlinear = lab @ _OKLAB_TO_LMS.T
-    lms = lms_nonlinear**3
-    return lms @ _LMS_TO_LRGB.T
+from pylette.src.types import ExtractionMethod
 
 
 @register(ExtractionMethod.OKLAB)
